Refactor search filters and solr highlighting
