+++ app/models/search_builder.rb
@@ -2,7 +2,7 @@
 class SearchBuilder < Blacklight::SearchBuilder
   include Blacklight::Solr::SearchBuilderBehavior
   include BlacklightAdvancedSearch::AdvancedSearchBuilder
-  include Gpp::LateNoticeFilters
+  include Gpp::SearchFilters
   self.default_processor_chain += [
     :add_advanced_parse_q_to_solr,
     :add_advanced_search_to_solr,
